// packages/types/tests/schema-tests-3.1/license-identifier.ts
// Generated from https://github.com/OAI/OpenAPI-Specification/tree/v3.1-dev/tests/schema/pass/license_identifier.yaml
// Do not edit by hand; regenerate instead.
import type { OpenAPIObject } from '../../src/v3.1'

export const doc: OpenAPIObject = {
  openapi: '3.1.0',
  info: {
    title: 'API',
    summary: 'My lovely API',
    version: '1.0.0',
    license: {
      name: 'Apache',
      identifier: 'Apache-2.0',
    },
  },
  components: {},
}