// packages/types/tests/schema-tests-3.1/minimal-paths.ts
// Generated from https://github.com/OAI/OpenAPI-Specification/tree/v3.1-dev/tests/schema/pass/minimal_paths.yaml
// Do not edit by hand; regenerate instead.
import type { OpenAPIObject } from '../../src/v3.1'

export const doc: OpenAPIObject = {
  openapi: '3.1.0',
  info: {
    title: 'API',
    version: '1.0.0',
  },
  paths: {},
}